+++ tests/django_test_app/models.py
@@ -14,3 +14,8 @@ class PartialVersionModel(models.Model):
     partial = semver_fields.VersionField(partial=True, verbose_name='partial version')
     optional = semver_fields.VersionField(verbose_name='optional version', blank=True, null=True)
     optional_spec = semver_fields.SpecField(verbose_name='optional spec', blank=True, null=True)
+
+
+class CoerceVersionModel(models.Model):
+    version = semver_fields.VersionField(verbose_name='my version', coerce=True)
+    partial = semver_fields.VersionField(verbose_name='partial version', coerce=True, partial=True)